fixing tests + changing numbers to long

+++ src/main/java/io/myproject/firstproject/springbootstarter/statistic/NumbersWrapper.java
@@ -3,13 +3,13 @@ package io.myproject.firstproject.springbootstarter.statistic;
 import java.util.List;
 
 public class NumbersWrapper {
-  private List<Integer> numbers;
+  private List<Long> numbers;
 
-  public List<Integer> getNumbers() {
+  public List<Long> getNumbers() {
     return numbers;
   }
 
-  public void setNumbers(List<Integer> numbers) {
+  public void setNumbers(List<Long> numbers) {
     this.numbers = numbers;
   }
 }
